upgrade renderer to use 32-bit integer math

+++ Source/r_plane.h
@@ -35,9 +35,9 @@
 #define PL_SKYFLAT (0x80000000)
 
 // Visplane related.
-extern  short *lastopening;
+extern  int *lastopening;
 
-extern short floorclip[], ceilingclip[];
+extern int floorclip[], ceilingclip[];
 extern fixed_t yslope[], distscale[];
 
 void R_InitPlanes(void);
